screenrecorder: Fix screen sharing on devices without built-in displays
For external displays, it seems like the button name is the name of the
monitor, as opposed to "Built-in display". This makes tests that share
the entire screen fail on Chromebox devices (e.g. fizz and puff).
We now just select the first display, ignoring the name. This logic is
also more similar to what's used for window and tab sharing.

BUG=chromium:1373880
TEST=CQ

Change-Id: I9b5b7dbb5e9530296af7211b09bf179911e0f4b9