dlp: Fix screen sharing on devices without built-in displays
For external displays, it seems like the button name is the name of the
monitor, as opposed to "Built-in display". This makes tests that share
the entire screen fail on Chromebox devices (e.g. fizz and puff).
We now just select the first display, ignoring the name. This logic is
also more similar to what's used for window and tab sharing.

BUG=chromium:1373880
TEST=tast run -build=true <ip>
dlp.DataLeakPreventionRulesListScreenshare*

Change-Id: I9b5b7dbb5e9530296af7211b09bf179911e0f4b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3990926
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Tested-by: Muhamed Parić <muhamedp@google.com>
Commit-Queue: Muhamed Parić <muhamedp@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>