crostini: Open launcher before waiting for icon load
crostini.Launcher test installs an app, and then ensures that the app
icon gets loaded before proceeding - it expects that the icon gets
requested for display in launcher, but never actually opens launcher to
display the icon, so there is no good reason for UI to actually request
the icon. The test passes because launcher zero-state app search provider
was running in background, updating results whenever set of installed
apps changed, and as a result was requesting an icon for zero-state
results. CL:3984753 changed that (i.e. zero-state search provider
stopped doing unnecessary work when launcher was hidden), and broke this
implicit test expectation. This updates the test to open the launcher
before waiting for icon load.

BUG=b:256233417
TEST=tast run nautilus crostini.Launcher*

Change-Id: I0783ab527b0c5c98bea4bfe28a3f01f4c948dda2