arc: Ensure ArcFileReaderTest is recognized by Chrome after installation
With CL:3934653, the Files app gets the list of apps to open a specific
file via AppService instead of directly asking ARC via mojo.
Since there can be a time lag between the app install completion on the
ARC side and the corresponding update of the app list on the Chrome
side, we should ensure that the app list displayed by the Files app is
updated before trying to open a test file with ArcFileReaderTest.

BUG=None
TEST=Run arc.MyFiles.vm etc. after inserting a delay (~15 secs) in the
TEST=app list update on the Chrome side. The tests succeed only when
TEST=patched with this change.

Change-Id: Ib0444f05e28541a869549e867150867165520f14