arc: Use "Open with..." context menu to open file with ArcFileReaderTest
With CL:3934653, the Files app gets the list of apps to open a file with
via Chrome's AppService instead of asking ARC via mojo. As a result, it
can take time for the Files app to recognize the ArcFileReaderTest app
after the app is installed via adb.

In tests that use ArcFileReaderTest, if the "Open" button is clicked
before the test app is recognized by the Files app, the test file is
unexpectedly opened with the default Text app. This is because it is
the only app that can open text files.

To avoid this and eusure the test file is opened with ArcFileReaderTest,
use the "Open with..." context menu instead of the "Open" button. The
context menu is available only after multiple apps are available to open
the test file, i.e., ArcFileReaderTest is recognized by the Files app.

Note that we do not call WaitUntilExists for the "Open with..." context
menu. This is because LeftClick (called from ClickContextMenuItem) waits
until the context menu appears and thus is sufficient.

BUG=None
TEST=Run arc.MyFiles.vm etc. after inserting a delay (~15 secs) in the
TEST=app list update on the Chrome side. The tests succeed only when
TEST=patched with this change.

Change-Id: Ib0444f05e28541a869549e867150867165520f14