network: add test for IPv6 overlay of IKEv2
This patch updated the tast test to check if shill can configure the
IPv6 address for the overlay IP of IKEv2. In crrev.com/c/3990073, the
client requests the server to send an overlay IPv4 address and IPv6 one
as virtual IP addresses. In this patch, the server supports dual stack
and checks if the overlay IP addresses can be configured by the client
by pinging.

BUG=none
TEST=tast run $DUT network.VPNConnect.*
TEST=../platform/tast/fast_build.sh -T

Change-Id: I7e69e2c54db84631adcf9266fe096dbc89b5ce97