tast-tests: Add lacros variant for filemanager.DragDrop
The existing test relies on a chrome app to create a drag drop handler.
Unfortunately chrome apps are deprecated, so this not only migrates the
chroem app to a PWA but also adds a LaCrOS variant of the test to ensure
drag drop from Files app registers in the lacros browser.

BUG=b:216707053
TEST=tast run filemanager.DragDrop*

Change-Id: I00f44dd037e0da6833897f82465c2b054f14182b