Clear WebAuthn session before starting test
We currently use a third-party WebAuthn site for testing WebAuthn
behavior, and it keeps session-id cookie at browser side. We navigate to
the /logout path during cleanup before, but it's better that tests don't
rely on this behavior. Move the logout logic to the start of the test
instead of during the cleanup stage, so that a new session is guaranteed
during test. This has a side benefit that the dumped ui_tree and the
chrome screenshot contain more meaningful content about the page now.

BUG=b:252923705
TEST=u2fd.Webauthn* hwsec.WebauthnU2fMode policy.WebauthnFactors

Change-Id: Ib3446fb85d7dfef51fdc487a09ab0f23db68a22b