tast: pmksa_caching: fix failure to roam back to initial AP
PMKSA Caching test connects to AP0, roams to AP1 expecting
EAP exchange not to be skipped and later deconfigures AP1,
expecting DUT to disconnect and roam back to AP0. EAP exchange
is expected to be skipped when roaming back to AP0. However,
this test can fail to fall back to AP0 if the device does not
disconnect and fallback to AP0 within 30 secs timeout period.
Explicitly, roam back to AP0 and ensure that EAP exchange is
skipped.

BUG=b:253111845
TEST=tast run ${HOST} wifi.PMKSACaching on Jacuzzi
TEST=Device managed to roam back to AP0 and EAP exchange was
     skipped.

Change-Id: I2d43bbb75db0d035c4d3953b2b9525e796a7a178