ui: Take screenshot of WebRTC Internals in MeetCUJ
MeetCUJ tests sometimes fail waiting for the download notification when
fetching the JSON dump from chrome://webrtc-internals. I have tried and
failed to reproduce this with crosfleet. Take a screenshot of the WebRTC
Internals window when the dumpWebRTCInternals function fails.

BUG=b:255343902
TEST=tast run <DUT> ui.MeetCUJ.lacros_2p with CL:3994032 and verify that
webrtc-internals.jpg is a screenshot of the WebRTC Internals window.

Change-Id: I3528a85f3184a53f8640ef8f5cef05f1b42e8af6