tast-tests: Deflake SelectAvatarFromFile
The operation set should be: 1) click the default avatar from options,
and 2) verify that the avatar changes. Now about 5% tests will fail
because 2) is not verified. Giving the whole operation set another
chance would very likely fix this issue.

I looked into the screenshots of this failure, and all of them happens
to be in dark mode. Turning this test explicitly into light mode will
also possibly fix this issue.

BUG=b:252945595
TEST=tast run $DUT personalization.SelectAvatarFromFile

Change-Id: I7b3fea5bd58b39687eabdf887117c45a4f0a19bf