Tast-test: Add services to stop in power setup
Add fwupd and avahi to the list of services to stop in
power setup as a part of noise cancellation. Default
behavior for both services are disabled.

BUG=b:254924845
TEST=tast run arc.PowerIdlePerf.noarc; Test passed and
service states are changed from 'running' to 'waiting'
in the middle of the test.

Change-Id: I822df7377c2903cf69590afa18d17b89a87cfaec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3994008
Tested-by: Yanye Li <yanyeli@google.com>
Auto-Submit: Yanye Li <yanyeli@google.com>
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Yanye Li <yanyeli@google.com>
Reviewed-by: Chenghao Piao <cpiao@google.com>