ARC Demo Mode: Add retry and mouse movement to increase success rate
The recently added ARC demo mode test would succeed on a higher-end
hatch DUT, but would mostly fail on other devices (esp. lower-end).

One disruption to the test is when some blocking statement excceeds 60
seconds and triggers demo session restart. One way to mitigate is to
periodically move the mouse to prevent idling for 60 seconds or more.

This CL does a few things in attempt to increase the success rate:
1. Relax the timeout value;
2. On a separate goroutine, periodically move mouse;
3. Add retry.

This CL also creates variants for vm and non-vm, to help identify the
source of potential issue, when the test fails in the future.

BUG=b:235602920
TEST=tast run dut arc.DemoMode

Change-Id: I3e57c464db00d900f2239e65d842a6ce6e740ee1