ui: Add custom screenshots to ScreenshotRecorder
Add the ability for tests to take screenshots at custom times during the
test. In making this change, I slightly adjusted what the default values
in NewScreenshotRecorder means. 0 max images for the recorder means that
we will not take any screenshot when Stop is called. A 0 interval means
we wil not take any interval screenshots. Thus, you could do
combinations like 0 max images and 0 interval to never take any
screenshots, and 1 max image and 0 interval to take only a single
screenshot at the end of the test.

These changes are to make the custom screenshot functionality robust,
and prevent any duplicate screenshots that would be taken. For example,
in GoogleSheetsCUJ, we close the Google Sheets window at the end of
recorder.Run, but before we stop the screenshot recorder. Thus, the
"end" screenshot would just be of the desktop. With custom screenshots,
GoogleSheetsCUJ now takes a screenshot manually before closing the
window, so that we can properly see the window state.

BUG=b:255577826
TEST=tast run <DUT> ui.GoogleSheetsCUJ ui.GoogleSlidesCUJ
ui.ArcYoutubeCUJ ui.TabSwitchCUJ, and ensure that the output directory
had a single screenshot with the "-custom" suffix.

Change-Id: Ibc07fb347a470374d7b727abc2f20a040d9296e8