ui: Add custom screenshots to ScreenshotRecorder
Add the ability for tests to take screenshots at custom times during
the test, including before and after recorder.Run. In making this
change, I slightly adjusted what the default values in
NewScreenshotRecorder means. 0 max images for the recorder means that
we will not take any screenshot when Stop is called, and 1 max image
will be a screenshot recorder that only takes an image at the end of
recorder.Run. Additionally, interval will be ignored when there is less
than 2 max images.

These changes are to make the custom screenshot functionality robust,
and prevent any duplicate screenshots that would be taken. For example,
in GoogleSheetsCUJ, we close the Google Sheets window at the end of
recorder.Run, but before we stop the screenshot recorder. Thus, the
"end" screenshot would just be of the desktop. With custom screenshots,
GoogleSheetsCUJ now takes a screenshot manually before closing the
window, so that we can properly see the window state.

BUG=b:255577826
TEST=tast run <DUT> ui.GoogleSheetsCUJ ui.GoogleSlidesCUJ
ui.ArcYoutubeCUJ ui.TabSwitchCUJ, and ensure that the output directory
has a single screenshot with the "-custom" suffix.

Change-Id: Ibc07fb347a470374d7b727abc2f20a040d9296e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3995703
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Tested-by: Navin Ramsaroop <ramsaroop@google.com>