tast-tests: Add a retry loop helper utility for flaky tests
In this change we're adding a new struct to encapsulate a retry loop in
a tast test which has stable/unstable variants.

Any test that needs to repeat certain tests need some common variables
to track total attempts, max attempts, whether retries are enabled, etc
and some helper methods to encapsulate the retries. With this utility we
can avoid code repetition in such tests.

I've updated only single test in this change to demonstrate usage of
this utility. More tests will be changed in a follow up.

BUG=b:256684807
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: I8681a8ffadf59effe6893e021ce3b25c45767c0d