cuj: Make ui.LoginPerf terminate on first error
ui.LoginPerf has several test scenarios and historically it did not
terminate on individual scenarios failures. But test is built
incrementally so right now the same error is very likely to appear on
all the runs.  This CL makes it fail early to highlight the issues. We
can rethink this after current issues are fixed.

BUG=b:237400719,b:218559327
TEST=tast run <dut> ui.LoginPerf*

Change-Id: Ie5242d589f16373555bbe3fc0f9ce0cc635f5d87