vpn: Switch Envs for VPN servers
This patch switches the virtualnet Env to set up the two VPN servers, so
that the first server is set up on the server env, and the second server
is on the router env. Since there is no default route from the router
env to the server env, the VPN server interface (overlay interface) will
no longer be reachable from DUT. This could avoid some unexpected
results in the tests.

BUG=None
TEST=fast_build.sh -T
TEST=tast run $DUT network.VPNConnect*

Change-Id: Ibd0e7214f4725841ad0f4d2f399360643c62f347