tast: use reliable multi word cases
Previously we updated the multi word suggestion test cases to use the
string "goo" to trigger a suggestion of "good morning". This example is
not reliably showing for successive attempts in the same input (due to
internal logic in the engine). A more reliable case is "hi there h" to
trigger "hi there how are you".

BUG=b:255679709
TEST=tast -verbose run -build=true 127.0.0.1:9222
inputs.PhysicalKeyboardMultiwordsuggestion

Change-Id: I042350c5ef32b0742863c577b095b24af6a6ecad