tast-tests: Refactor EmulatedBTPeerDevice to better prepare device
Testing during pairing test development found that the btpeers
configured as an EmulatedBTPeerDevice were not adequately
initialized as the desired btpeer device type. This change updates
the initialization steps to include those that are necessary and
removes steps that are not (determined by reviewing Chameleond
source code), adds support to specify pairing auth methods (with new
defaults that match those in Tauto so pin code pairing is not
necessary), and refactors the NewEmulatedBTPeerDevice to accept
config options as a struct that is easier to use.

With these changes, all test writers need to do to configure a btpeer
and make it ready for pairing is call NewEmulatedBTPeerDevice. The
act of completing the paring from the DUT end is then up to the test.
There is no cleanup needed within the test either, the fixture's reset
is sufficient for clearing the btpeer emulation.

BUG=b:217633791
TEST=tast unit tests passed; passed the following remote bluetooth
tast tests:
bluetooth.TestConnectToBTPeers
bluetooth.DiscoverEmulatedBTPeerDevices

Change-Id: I87e4ba37f4db24b0dc8397aed73624cb2a3b0a39