network: Add StableShillEAPSlot test
The underlying PKCS#11 slot ID is not stable, meaning that it is
incorrect to assume that system and user slot always use slot ID
0 and 1 respectively. On specific cases, these values might be
changed.

This patch adds the test to ensure shill can correct itself. This
is done by:
1. Loading a system and user slot (normally 0 and 1).
2. Pushing certificate and key to both of the slots.
3. Stopping shill (triggers shill's Save())
4. Re-order the slots, by adding a temporary token on slot 0
   (normally the state would be system slot 2 and user slot 0).
5. Re-start shill (triggers shill's Load())
6. Observe that the new loaded certificate and key ID are valid.

On this patch, Chrome needs to be restarted in order for
NetworkCertMigrator to know of the latest state. Otherwise, it
will remove the property.

Also see: go/cros-stable-cert-id

BUG=b:255445416
TEST=tast run <DUT> network.StableShillEAPSlot;

Change-Id: I9477f964bb1bf2cb1dacae4d5eb8935287144127