camera: Add a third state for PrivacySwitch
Some camera modules have support for a Privacy GPIO, but the OEM might
decide to not expose that pin to the user.

Flag this kind of modules by setting `has-privacy-switch = False` on the
`cros config`.

Because a model can be shipped with different camera modules, when we
detect `has-privacy-switch` as `False` ignore the content of the v4l2
controls.

BUG=b:256996604
TEST=checked on lazor and eldrid

Change-Id: I1ea5cc88ae75f218d2bb7d1e0f14f9531c1984a1