tast-tests: Adjust install timeout for enterprise.ARCProvisioning
Installation in Play Store can take up to 10 minutes due to hygiene job
but currently we're taking the entire test timeout of 15 minutes for
waiting for install to finish. This prevents the second retry from
executing due to not enough time being left for it to run.

In this change, we're going to limit waiting for app installation to 10
mins only. Leaving another 5 minutes for second attempt to succeed.

BUG=b:254838300
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: If3b49e72c6314bb8159b4b117ad930ee3ca7e069
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4000429
Reviewed-by: Yao Li <yaohuali@google.com>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>