Tast-test: Add multicast disabler to power setup
Add multicast disabler through "ifconfig" to power setup
structure and the default behavior is to disable them on
ethernet and wlan interfaces. Multicast will be re-enabled
following the exit of the tests.

BUG=b:254924845
TEST=tast run arc.PowerIdlePerf.noarc; Test passed and the
log shows multicast disabled/re-enabled. This is also
verified by "ifconfig | grep flags" in the middle of the
test. Tested under both wlan and ethernet connections.

Change-Id: I172922e83052e6219d8fa93e03e0598a4c5840d0