tast-tests: Update NewTestFixture's default router/pcap to UnknownT
In some tests, we invoke NewTestFixture which currently defaults to a
legacy router type for the router and pcap. This means if a new router
is generated, the router resolver in newRouter() won't trigger since the
default router type isn't support.UnknownT. We change the default types
to UnknownT so the resolver is triggered in this base case. In other
cases where the support is specified, the behavior isn't changed.

BUG=b:257334215,b:250673442
TEST=tast run wifi_matfunc with this patch on brya against openwrt
router/pcap and gale router/pcap

Change-Id: Iecdaaa2803671a48aab58faced7e3ac2a03c8578