tast-tests: Update newRouter to always infer router
In some tests, we invoke NewTestFixture which currently defaults to a
legacy router type for the router and pcap. This means if a new router
is generated, the router resolver in newRouter() won't trigger since the
default router type isn't support.UnknownT. We force the router to
always resolve the router type. We throw an error if we cannot resolve
the router type. We also change the default types
to UnknownT instead of defaulting to LegacyT. For future router
adapters, we will have to update the router resolving mechanism

BUG=b:257334215,b:250673442
TEST=tast run wifi_matfunc with this patch on brya against openwrt
router/pcap and gale router/pcap

Change-Id: Iecdaaa2803671a48aab58faced7e3ac2a03c8578