tast-tests: Ensure Play Store has an app visible in enterprise tests
Multiple enterprise tests are using a helper to validate that Play Store
app catalog is not empty but it is only checking absence of empty
message and not validating whether an app is actually visible.

In this change we're going to ensure there is at least one app visible.

BUG=b:252934524
TEST=tast run DUT enterprise.ARCAvailableAppInstall.vm
TEST=tast run DUT enterprise.ARCBlockedAppInstall.vm
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: Id38b3483d625beab19562c7af381b01939ac4b48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4005069
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Yao Li <yaohuali@google.com>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>