Verify in demomode.SWA that the splash screen gets removed
The splash screen is a barebones wallpaper that temporarily covers the
screen while the Demo Mode session is starting. This test needs to wait
until the splash screen is removed (after a timer) before we can move
the mouse to get the app to go from fullscreen to windowed mode.

This issue didn't manifest at first because the SWA used to take longer
to launch on the first session after setup, because it was downloading
the content. Now that the content is downloaded during the setup flow,
the SWA launches after a couple of seconds, before the splash screen
closes.

BUG=b:249608648
TEST=tast run <device-ip> demomode.SWA

Change-Id: Ie0280848b3bb8d328fcbd3ef34ada2a9c87385ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4009420
Tested-by: Jackson Tadie <jacksontadie@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Jackson Tadie <jacksontadie@google.com>
Reviewed-by: Yves Arrouye <drcrash@chromium.org>