tast-tests: Improve local development for tast-tests
This change adds two runtime flags for CUJ tast tests: cuj.isLocal and
cujrecorder.keepWifi.

cuj.isLocal quickens the runtime of tests by skipping CPU idle and
cooldown time. This flag is useful for quickly checking on changes of a
tast test, when performance data is not important.

cujrecorder.keepWifi forces the cujrecorder.Recorder to keep the Wifi in
its initial state. Devices that require Wifi for an active SSH
connection will fail CUJ tests, because we turn the Wifi off by default.

BUG=b:256715411
TEST=tast run -var=cuj.isLocal=true -var=cujrecorder.keepWifi=true
<DUT> ui.QuickCheckCUJ

Change-Id: I1e2684282e2d7dc51859ef316d5313f79daa065e