tast-tests: Add ash.WaitUntilDesksFinishAnimating
Add a function that waits until desks finish animating, to help tests
properly wait without the need for testing.Sleep. This change removes
the sleep in ui.DesksTrackpadSwipePerf that waits for desks to finish
animating. Additionally, this change updates ui.DesksCUJ to properly
wait for the desks animation to finish, and separately wait for the
active desk to stabilize.

BUG=b:246782864
TEST=tast run <DUT> ui.DesksCUJ ui.DesksTrackpadSwipePerf

Change-Id: I8715c64f622a7a13d9e7cb0fa8eac3d8e75f0175