tast-tests: Drop the AddSessionMount test
After doing some analysis, this test seems to be fairly obsolete. It's
testing using AuthSession+Mount to create and mount users in a couple
different configurations. However, this particular combination of
old+new APIs is now obsolete.

The test could be rewritten to only use the new APIs, but then what it's
doing would end up overlapping with other new tests which already do
this.

BUG=b:258042583
TEST=fast_build.sh -T

Change-Id: Idc85c3a84c0e5fe14632da42ede4a01731907974