printer: Fix print job verification for printer.Print test
printer.Print will use the print management dialog to determine if
printing was successful by looking to see if there is a print job.  If
a test runs prior to printer.Print that has a print job in the history,
printer.Print will not work since multiple nodes will match.  Fix this
by clearing the history before we perform any printing for any tests
that do verification via the print management dialog.

BUG=b:257296435
TEST=tast run dut printer.PrintExtension.complete printer.Print
TEST=tast run dut printer.PrintExtension.lacros_complete printer.Print.lacros
TEST=tast run dut printer.PrintExtension.lacros_complete policy.Printers

Change-Id: I076f489cfa01fadbd7a4c4eadae1cbeb6e568095