tast-tests: Add base64 param support to XMLRPC client
Python XMLRPC, which is what our XMLRPC client is for, supports a
base64 type (https://docs.python.org/3/library/xmlrpc.client.html).
This change adds support for both using it as a parameter in requests
and parsing it from responses. From a user's perspective, both are
seen in the decoded form, []byte. The XMLRPC client unit tests have
also been updated to validate this support.

This support was added to allow for calling Chameleond XMLRPC methods
that require a base64 param, and this support was confirmed to work
with these methods in testing.

BUG=b:217633791
TEST=tast unit tests pass, including updated ones

Change-Id: I34a4ca29202322d4affa50354073523b7a1c1c02