tast-tests: Add FastPairInitialPair and FastPairSubsequentPair tests
WIP - Still needs UI automation, currently relies on manual click.

This change corrects the parameters on the Chameleond fast pair RPC
endpoints, adds a new fixture for fast pair testing, and adds two
new fast pair tests, FastPairInitialPair and FastPairSubsequentPair,
that use these features. While the tests do not pass, they do seem to
be preforming their tests correctly. As they are not expected to pass,
they have not been added to any test suite and are just for manual
testing only for debugging a possible bug with DUT and/or btpeer fast
pair support.

FastPairInitialPair observation:
Can see the fast pair notification appear, clicking on connect causes
a connecting progress bar to appear, then the notification goes away.
It does not seem to pair or connect to the btpeer.

FastPairSubsequentPair observation:
Can see the fast pair notification appear, clicking on connect causes
a connecting progress bar to appear, then after a while it says it
failed to connect. It does not seem to pair or connect to the btpeer.

BUG=None
TEST=tast unit tests pass; new tests run as expected but do not pass

Change-Id: I7ff0319672ddabd8af419611c987d0dbfc97feab