cellular: Check for test EUICC when rebooting modem
This avoids erroneously rebooting the cellular modem when the device has
a test EUICC present but no valid SIM, as tests that utilize a test EUICC
don't require a valid SIM to succeed. It will still attempt to reset the
modem if none is exported but it won't return an error if the modem
exists but there is no valid SIM after rebooting.

BUG=b:245795270
TEST=tast run <DUT> '("group:cellular" && "cellular_sim_test_esim" && !"cellular_unstable
TEST=tast run <DUT> '("group:cellular" && "cellular_sim_active" && "cellular_unstable")'

Change-Id: I8134208fe3975efb8bcf09c3de7d691ad6a5c38e