tast-tests: Drop the EphemeralAuthSession test
With the changes made in crrev.com/c/3989744 this test is now largely
redundant. The newly rewritten version of AuthSessionUnlock will set up
a user and do login and verification, and is parameterized to cover both
persistent and ephemeral users.

Since this test is also using API features that are now obsolete
(CheckVault, AuthSession without AuthFactor) we want to turn it down in
favor of the one that covers the modern equivalent.

BUG=b:258042583
TEST=fast_build.sh -T

Cq-Depend: chromium:3989744
Change-Id: Ib612e71906005d4b163c61ce296b857dfdc5b16b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4010809
Reviewed-by: Joseph Pang <pjlee@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>