tast-tests: Simplify the kiosk tests
There are currently many (six!) different test files for kiosk
credentials, most of which are basically doing a bunch of different
variations of the same thing.

This change deletes most of them (particularly the ones focusing on
exercising the obsolete APIs) and consolidates the rest of the coverage
into two tests, one for ephemeral setups and one for persistent setups.

BUG=b:258042583
TEST=tast run <dut> cryptohome.KioskEphemeralMount.*
TEST=tast run <dut> cryptohome.KioskPersistentMount.*

Change-Id: I48557186ce2b15c7ef8bee15fb0b9c18dd3621f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4014549
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: David Welling <davidwelling@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>