tast-tests: Add a helper to poll Play Store app page
Simplified the polling of Play Store app detail page by introducing a
helper and using it in multiple tests with similar code.

Also, increased the reliability of polling by re-creating the
UIAutomator connection as it gets uninstalled when the policy changes.

Also, updated EnsurePlayStoreState helper to close/re-open Play Store
if the desired state is not reached in the inner loop.

BUG=b:256684807
TEST=tast run DUT enterprise.ARC*.vm

Change-Id: I986af58c04b5d634a6a9e837be9901ea2d8462c9