faft: Improve on BatteryCharging
Sometimes even though setting the power state was
successful, the charger state did not change as
expected. Based on Stainless logs, this was the
case mostly on DUTs with type-A servos(rpm). Before
exiting the test, log usb port status. With this info,
we might be able to use ec console command to manipulate
(shut) charger from the dut side on the port that
receives power.

BUG=b:200305302
TEST=manual on juniper, apel, and hayato

Change-Id: I5209edf1736dc317a26f935954e3a4e3d1cf3e90