policyutil: Check global timeout in the enrolled fixture
At the moment the tests using the enrolled fixture are at the end of
shard 1 of the informational tests, and there is not enough time to
execute the SetUp function of the enrolled fixture.
Adding a check to make the error message reflect the actual issue.

BUG=b:258159879
TEST=tast run -timeout=120 <DUT> policy.AllowDinosaurEasterEggEnrolled

Change-Id: If0b19e5fa48d12cde4e187812dd418367415c4b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4016573
Tested-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Gabor Zsolt Magda <gabormagda@google.com>