tast-tests: Fix the deferred disconnection in wifi.RoamFT
The RoamFT test put the deferred disconnection before the deferred
teardown of ap1. If the roaming was successful, the DUT ended up
connected to ap1 at the exit of runOnce(), the DUT received a
deauthentication frame from ap1 and got deauthed first and the
subsequent deferred disconnect failed. crrev/c/3232168 added a second
deferred disconnection after the deferred teardown of ap1 to make sure
the DUT was disconnected before the teardown of ap1, and disabled the
first deferred disconnection if the roaming was successful. However,
the second deferred disconnection was also disabled if the roaming was
successful, so that none of these two deferred disconnections was
called, leaving the DUT connected to the ap1. Most devices disconnect
from ap1 immediately after ap1 sending the deauthentication frame at the
exit of the first runOnce(); while some devices stay connected and
therefore the scan in the second runOnce() is too long so that the
connection to ap0 fails. net.log shows that the DUT disconnects after
tearing down ap1, for example,
WiFi wlan0 roamed to Endpoint ce:69:3c:4a:f1:d3
Stopping hostapd
wlan0: CTRL-EVENT-DISCONNECTED bssid=ce:69:3c:4a:f1:d3 reason=4

Remove the condition of roamSucceeded before the second deferred
disconnection. If the roaming fails, the DUT is connected to ap0 at the
exit of runOnce(), the first deferred disconnection is called before
tearing down ap0; otherwise the DUT is connected to ap1 and the second
deferred disconnection is called before tearing down ap1. The DUT is
guaranteed to disconnect from the AP it connects to before tearing down
this AP.

BUG=b:253519667, b:231669774
TEST=tast run $DUT  wifi.RoamFT.mixed_psk
The tast log shows
Wait for the service to be idle
Disconnected
The net.log shows
[service.cc(401)] Disconnecting from wifi_psk_4: D-Bus RPC

Change-Id: I42095119ff877642da66f56bc0a55f7c9175dbd6