tast-tests: Fix the defer function CleanDisconnectWifi() in wifi.RoamFT
In wifi.RoamFT, defer function tf.CleanDisconnectWifi(ctx) is called if
roamSucceeded is false. However, if the roaming is successful, it is
never called, leaving the DUT connected to the second AP. Most devices
disconnect from the second AP immediately after the second AP sending
deauth frame at the exit of the first runOnce(); while some devices stay
connected and therefore the scan in the second runOnce() is too long so
that the connection to the first AP fails.

Remove the condition of roamSucceeded before the second defer function
tf.CleanDisconnectWifi(ctx). If the roaming fails,  the first defer
function tf.CleanDisconnectWifi(ctx) is called before tearing down the
first AP; otherwise the second defer function
tf.CleanDisconnectWifi(ctx) is called before tearing down the second AP.
The DUT is guaranteed to disconnect from the AP at the exit of each
runOnce().

BUG=b:253519667
TEST=tast run $DUT  wifi.RoamFT.mixed_psk
The tast log shows
Wait for the service to be idle
Disconnected
The net.log shows
[service.cc(401)] Disconnecting from wifi_psk_4: D-Bus RPC

Change-Id: I42095119ff877642da66f56bc0a55f7c9175dbd6