[QuickStart] Update crossDeviceNoSignIn to be able to load in OOBE screens
This is an iteration of crossDeviceNoSignIn to prepare the fixture to
actually be able to start the tests in OOBE Welcome and allow deferred
login that will be happening as a result of tests.

Caveat: This requires a small change to chromium to add Quick Start
Screen to the wizard_controller, which I will wait on OOBE team for
when it's ready. This test can be run manually by deploying the
small change (crrev/c/4018885) locally to the DUT.

BUG=b:257332394
TEST=Ran a tast test flow loading in the QuickStart OOBE screen

Change-Id: I6bca89e96549541123ef6dac42b88d0d452803d7