tast-tests: Add multi-user verify to the unlock tests
This adds support for testing multiple simultaneous users in the
AuthSessionUnlock test. This is a test case that we have very limited
tast coverage for.

The change also breaks the test up into persistent and ephemeral
versions. Adding multiple users (not supported in the ephemeral case)
was just increasing the amount of "if ephermal" branches in the test too
much, so splitting the test into two made sense. This is already a
pattern we already use in a couple other places, for the same reason.

BUG=b:255985000
TEST=tast run <dut> cryptohome.AuthSessionUnlock.*
TEST=tast run <dut> cryptohome.AuthSessionUnlockEphemeral.*

Change-Id: Icc4e39bab84fb80d409deda3de83dcf0b44cf683
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4021390
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>