tast-tests: Rename, move and simplify arc.EnterpriseLogin
Most enterprise test cases are in the enterprise package so moving
arc.EnterpriseLogin to enterprise package. Also renamed it to ARCBoot
because the test is only verifying that ARC boots when enabled by policy
and does not boot when it is disabled.

We're also using a fake policy server in this test to avoid flakiness
due to connection with DM Server.

BUG=b:235602184
TEST=tast run DUT enterprise.ARCBoot*vm

Change-Id: I30eb480e0ab3bfbac6e9b7bbd02555a83b464fc3