power: Retry HidWake suspend/resume on servo keyboard reconnect
The HidWake test uses the keyboard emulator on a servo to wake the DUT
using a HID event. When the DUT resumes, it will also resume the
keyboard emulator which should be in a suspend state. Sometimes, this
process encounters an IO error which causes the test to be flaky. After
the servo keyboard reconnects, the wake count is reset to 0. This causes
the test to fail either the wake count check, or the wake source check
from powerd_dbus_suspend logging which relies on the wake source count.
The peripheral resume behavior of the keyboard emulator is not within
the scope of this test. And should not cause a test failure. This CL
updates the test to monitor for keyboard disconnects/reconnects and
retry suspending/resuming a limited number of times if it hits a
keyboard disconnect.

BUG=b:255825993, b:252890504
TEST=tast unit test passed (fast_build.sh -T). Changed condition to
trigger suspend/resume retries and saw the DUT suspend/resume multiple
times, as expected.

Change-Id: I60755f5e213288cd798a4548346b0021fdaaea29