ui: Properly check for tab crashes in MeetCUJ
This change fixes the issue where the tab crash checker was only
checking crashed tabs on Ash, which meant crashes on Lacros would never
have been caught. Additionally, we now properly check for tab crashes
both after the Meet call is loaded, after recorder.Run completes, and
also after the entire test is completed (using a defer call), so we can
be certain that the logs will properly show if some tab crashed during
the test. Lastly, this change updates the UMA metrics that are checked
to verify if a tab has crashed, because the previous set of metrics were
removed in crrev/c/3523209.

BUG=b:258029099
TEST=tast run <DUT> ui.MeetCUJ.lacros_16p on elm/elm devices. Elm boards
have an issue where Meet crashes on load every time, so this change
could properly be verified using it.

Change-Id: Ib1088bd3ba5e009b36e448f35cc9cfbf051b93b6