tast-tests: Extend polling time limit in arc.GhostWindow.fixup_r
This CL extends the time limit of a polling in arc.GhostWindow.fixup_r
from 10 seconds to 1 minute, since 10 seconds appeared to be too short
for the majority of boards used in the tast test. The polling waits for
the Play store data directory in the SDCard partition to be created.

BUG=b:186767881
TEST=arc.GhostWindow.fixup_r passed on octopus (apel) 10/10 times.

Change-Id: I2bfa54400b96cc725a59a393c12030246f4e4492